requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
ale-py==0.8.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.1
certifi==2022.12.7
click==8.1.6
cloudpickle==2.2.1
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
datasets==2.13.1
dill==0.3.6
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.6.1
fairscale==0.4.13
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
gym==0.26.2
gym-notices==0.0.8
h5py==3.9.0
huggingface-hub==0.16.4
importlib-resources==6.0.0
Jinja2==3.1.2
kaggle==1.5.16
keras==2.13.1
kiwisolver==1.4.4
labml==0.4.167
labml-helpers==0.4.89
labml-nn==0.4.135
libclang==16.0.6
lit==15.0.7
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
mkl-fft==1.3.6
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu12==12.2.4.5
nvidia-cuda-nvrtc-cu12==12.2.128
nvidia-cuda-runtime-cu12==12.2.128
nvidia-cudnn-cu12==8.9.4.25
nvidia-dali-cuda110==1.28.0
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
pandas==2.0.3
pathtools==0.1.2
Pillow==9.3.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soupsieve==2.5
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-logger==0.1.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorrt==8.6.1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
text-unidecode==1.3
timm==0.3.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchdata==0.6.1
torchtext==0.15.2
torchvision==0.15.2+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wandb==0.15.5
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
